stereoset: set label dtype to avoid issues with torch

+++ data_loader/stereoset.py
@@ -71,8 +71,8 @@ class StereoSet(CustomDataset):
         # every sample has 3 sentences with different labels
         # there is only the validation set
         n_sent = len(ds['validation']) * 3
-        g_val = np.zeros((n_sent, n_groups))
-        y_val = np.zeros((n_sent, 1))
+        g_val = np.zeros((n_sent, n_groups), dtype=int)
+        y_val = np.zeros((n_sent, 1), dtype=int)
         sent_val = []
 
         for i, sample in enumerate(ds['validation']):
